Kenzie/MAF-13: Updated notification logic.
