+++ app/src/main/java/com.terracode.blueharvest/listeners/homeListeners/RecordButtonListener.kt
@@ -3,26 +3,14 @@ package com.terracode.blueharvest.listeners.homeListeners
 import android.view.View
 import android.view.View.OnClickListener
 import com.terracode.blueharvest.HomeActivity
-import com.terracode.blueharvest.utils.NotificationTypes
+import com.terracode.blueharvest.utils.Notifications
 import com.terracode.blueharvest.utils.PreferenceManager
-import com.terracode.blueharvest.utils.Notification
-import java.time.Instant
-import java.time.format.DateTimeFormatter
 
 class RecordButtonListener(private val activity: HomeActivity) : OnClickListener {
 
     override fun onClick(view: View?) {
-        val startRecordingNotification = Notification(
-            NotificationTypes.NOTIFICATION,
-            "Started Recording Data",
-            DateTimeFormatter.ISO_INSTANT.format(Instant.now()).toString()
-        )
-
-        val stopRecordingNotification = Notification(
-            NotificationTypes.NOTIFICATION,
-            "Stopped Recording Data",
-            DateTimeFormatter.ISO_INSTANT.format(Instant.now()).toString()
-        )
+        val startRecordNotification = Notifications.getStartRecordingNotification()
+        val stopRecordNotification = Notifications.getStopRecordingNotification()
 
         // Get the current status from SharedPreferences
         val currentStatus = PreferenceManager.getRecordButtonStatus()
@@ -36,10 +24,10 @@ class RecordButtonListener(private val activity: HomeActivity) : OnClickListener
         // Optionally, you can perform additional actions based on the new status
         if (newStatus) {
             // If the status is true, record and send notification
-            PreferenceManager.setNotification(startRecordingNotification)
+            PreferenceManager.setNotification(startRecordNotification)
         } else {
             // If the status is false, stop recording and send notification
-            PreferenceManager.setNotification(stopRecordingNotification)
+            PreferenceManager.setNotification(stopRecordNotification)
         }
     }
 }
