--- app/src/main/java/com.terracode.blueharvest/utils/NotificationObject.kt
@@ -1,8 +1,0 @@
-package com.terracode.blueharvest.utils
-
-
-data class Notification(
-    val type: NotificationTypes,
-    val message: String,
-    val timestamp: String
-)
