+++ app/src/main/java/com.terracode.blueharvest/utils/constants/NotificationObject.kt
@@ -1,0 +1,10 @@
+package com.terracode.blueharvest.utils.constants
+
+import com.terracode.blueharvest.utils.constants.NotificationTypes
+
+
+data class Notification(
+    val type: NotificationTypes,
+    val message: String,
+    val timestamp: String
+)
